requirements.txt
einops==0.3.2
hydra-core==1.1.1
protobuf==3.20.*
psutil==5.8.0
pygame==2.1.2
tqdm==4.62.3
wandb==0.12.7
gsutil==4.51
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.2.3
gym==0.19.0
atari-py==0.2.5
scipy
torch
torchvision